backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
protobuf>=4.25.3
python-dotenv>=1.0.0
pydantic-settings>=2.2.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
llama-cpp-python>=0.2.0
psutil>=5.9.0
